// src/lib/api/generated/models/GameTypeResponse.ts
/* tslint:disable */
/* eslint-disable */
/**
 * PolystirolHub Backend
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: 0.1.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { exists, mapValues } from '../runtime';
/**
 * 
 * @export
 * @interface GameTypeResponse
 */
export interface GameTypeResponse {
    /**
     * 
     * @type {any}
     * @memberof GameTypeResponse
     */
    name: any | null;
    /**
     * 
     * @type {any}
     * @memberof GameTypeResponse
     */
    id: any | null;
    /**
     * 
     * @type {any}
     * @memberof GameTypeResponse
     */
    createdAt: any | null;
}

/**
 * Check if a given object implements the GameTypeResponse interface.
 */
export function instanceOfGameTypeResponse(value: object): boolean {
    let isInstance = true;
    isInstance = isInstance && "name" in value;
    isInstance = isInstance && "id" in value;
    isInstance = isInstance && "createdAt" in value;

    return isInstance;
}

export function GameTypeResponseFromJSON(json: any): GameTypeResponse {
    return GameTypeResponseFromJSONTyped(json, false);
}

export function GameTypeResponseFromJSONTyped(json: any, ignoreDiscriminator: boolean): GameTypeResponse {
    if ((json === undefined) || (json === null)) {
        return json;
    }
    return {
        
        'name': json['name'],
        'id': json['id'],
        'createdAt': json['created_at'],
    };
}

export function GameTypeResponseToJSON(value?: GameTypeResponse | null): any {
    if (value === undefined) {
        return undefined;
    }
    if (value === null) {
        return null;
    }
    return {
        
        'name': value.name,
        'id': value.id,
        'created_at': value.createdAt,
    };
}